requirements.txt
Jinja2==3.0.1
black==20.8b1
mypy==0.790
pylint==2.6.0
pyyaml==5.3.1
pytest==6.1.2
pytest-cov==2.10.1
